+++ src/ff8/battle/effects.h
@@ -1,0 +1,38 @@
+/****************************************************************************/
+//    Copyright (C) 2009 Aali132                                            //
+//    Copyright (C) 2018 quantumpencil                                      //
+//    Copyright (C) 2018 Maxime Bacoux                                      //
+//    Copyright (C) 2020 Chris Rizzitello                                   //
+//    Copyright (C) 2020 John Pritchard                                     //
+//    Copyright (C) 2024 myst6re                                            //
+//    Copyright (C) 2024 Julian Xhokaxhiu                                   //
+//    Copyright (C) 2023 Cosmos                                             //
+//    Copyright (C) 2023 Tang-Tang Zhou                                     //
+//                                                                          //
+//    This file is part of FFNx                                             //
+//                                                                          //
+//    FFNx is free software: you can redistribute it and/or modify          //
+//    it under the terms of the GNU General Public License as published by  //
+//    the Free Software Foundation, either version 3 of the License         //
+//                                                                          //
+//    FFNx is distributed in the hope that it will be useful,               //
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of        //
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         //
+//    GNU General Public License for more details.                          //
+/****************************************************************************/
+
+#pragma once
+
+namespace FF8BattleEffect {
+    enum Effect {
+        Leviathan = 5,
+        Quezacotl = 115
+    };
+}
+
+namespace FF8BattleEffectOpcode {
+    enum Opcode {
+        UploadTexture39 = 39,
+        UploadPalette75 = 75
+    };
+}